refactor: Merge recipes services into one file.
- Collocate favoriteRecipes API request into recipes service as it make
sense for them to be together, also makes it easier to maintain.
